requirements.txt
speechbrain==1.0.2
torchmetrics
numpy==1.26.4
tqdm
matplotlib
pandas
wandb
comet_ml
hydra-core
librosa
ipykernel
torchvision

black
isort
pre-commit
